Let GCC check arguments of some more printf-style functions, also for plugins and codecs.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@12815 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/battery_bench.c
@@ -459,7 +459,7 @@ int main(void)
 #endif
                 "\n"
                 ,BATTERY_LOG,rb->global_settings->battery_capacity,
-                BUF_SIZE / sizeof(struct batt_info));
+                BUF_SIZE / (unsigned)sizeof(struct batt_info));
             rb->close(fd);
         }
         else
